Replace ironic deployment with IrSO in e2e basic

+++ hack/ci-e2e.sh
@@ -33,7 +33,7 @@ echo "BMO_E2E_EMULATOR=${BMO_E2E_EMULATOR}"
 export E2E_CONF_FILE="${REPO_ROOT}/test/e2e/config/ironic.yaml"
 export E2E_BMCS_CONF_FILE="${REPO_ROOT}/test/e2e/config/bmcs-${BMC_PROTOCOL}.yaml"
 
- case "${GINKGO_FOCUS,,}" in
+case "${GINKGO_FOCUS,,}" in
   *upgrade*)
     export DEPLOY_IRONIC="false"
     export DEPLOY_BMO="false"
@@ -59,20 +59,27 @@ sudo apt-get install -y libvirt-dev pkg-config
 # Build the container image with e2e tag (used in tests)
 IMG=quay.io/metal3-io/baremetal-operator:e2e make docker
 
-virsh -c qemu:///system net-define "${REPO_ROOT}/hack/e2e/net.xml"
-virsh -c qemu:///system net-start baremetal-e2e
+if ! sudo virsh net-list --all | grep baremetal-e2e; then
+    virsh -c qemu:///system net-define "${REPO_ROOT}/hack/e2e/net.xml"
+    virsh -c qemu:///system net-start baremetal-e2e
+fi
 
 # Allow traffic between docker bridges and the metal3 interface
 sudo iptables -I FORWARD -i br-+ -o metal3 -j ACCEPT
 sudo iptables -I FORWARD -i metal3 -o br-+ -j ACCEPT
 sudo iptables -L FORWARD -n -v
 
-# This IP is defined by the network we created above.
-IP_ADDRESS="192.168.222.1"
-
-# This IP is also defined by the network above, and is used consistently in all of
-# our e2e overlays
-export IRONIC_PROVISIONING_IP="${IP_ADDRESS}"
+if [[ "${GINKGO_FOCUS}" == "basic" ]]; then
+    # This IP address is used to reach virtual BMCs. The basic test is using
+    # irso, and vbmc container is running on host, so it can be reached through
+    # the docker network from kind node. Hence we need to configure vbmc to
+    # listen to docker network IP address instead of metal3 interface IP.
+    # Only relevant with ipmi, redfish will listen to every interface.
+    IP_ADDRESS="172.18.0.1"
+else
+    # This IP is defined by the network we created above.
+    IP_ADDRESS="192.168.222.1"
+fi
 
 # Build vbmctl
 make build-vbmctl
@@ -81,7 +88,7 @@ make build-vbmctl
 
 if [[ "${BMO_E2E_EMULATOR}" == "vbmc" ]]; then
   # Start VBMC
-  docker run --name vbmc --network host -d \
+  docker start vbmc || docker run --name vbmc --network host -d \
     -v /var/run/libvirt/libvirt-sock:/var/run/libvirt/libvirt-sock \
     -v /var/run/libvirt/libvirt-sock-ro:/var/run/libvirt/libvirt-sock-ro \
     quay.io/metal3-io/vbmc
@@ -98,7 +105,7 @@ elif [[ "${BMO_E2E_EMULATOR}" == "sushy-tools" ]]; then
   # Sushy-tools variables
   SUSHY_EMULATOR_FILE="${REPO_ROOT}"/test/e2e/sushy-tools/sushy-emulator.conf
   # Start sushy-tools
-  docker run --name sushy-tools -d --network host \
+  docker start sushy-tools || docker run --name sushy-tools -d --network host \
     -v "${SUSHY_EMULATOR_FILE}":/etc/sushy/sushy-emulator.conf:Z \
     -v /var/run/libvirt:/var/run/libvirt:Z \
     -e SUSHY_EMULATOR_CONFIG=/etc/sushy/sushy-emulator.conf \
@@ -118,29 +125,33 @@ IMAGE_DIR="${REPO_ROOT}/test/e2e/images"
 mkdir -p "${IMAGE_DIR}"
 
 ## Download disk images
-wget --quiet -P "${IMAGE_DIR}/" https://artifactory.nordix.org/artifactory/metal3/images/iso/"${IMAGE_FILE}"
-wget --quiet -P "${IMAGE_DIR}/" https://artifactory.nordix.org/artifactory/metal3/images/sysrescue/systemrescue-11.00-amd64.iso
+if [[ ! -f "${IMAGE_DIR}/${IMAGE_FILE}" ]]; then
+    wget --quiet -P "${IMAGE_DIR}/" https://artifactory.nordix.org/artifactory/metal3/images/iso/"${IMAGE_FILE}"
+    wget --quiet -P "${IMAGE_DIR}/" https://artifactory.nordix.org/artifactory/metal3/images/sysrescue/systemrescue-11.00-amd64.iso
+fi
 
 ## Start the image server
-docker run --name image-server-e2e -d \
+docker start image-server-e2e || docker run --name image-server-e2e -d \
   -p 80:8080 \
   -v "${IMAGE_DIR}:/usr/share/nginx/html" nginxinc/nginx-unprivileged
 
 # Generate ssh key pair for verifying provisioned BMHs
-ssh-keygen -t ed25519 -f "${IMAGE_DIR}/ssh_testkey" -q -N ""
+if [[ ! -f "${IMAGE_DIR}/ssh_testkey" ]]; then
+    ssh-keygen -t ed25519 -f "${IMAGE_DIR}/ssh_testkey" -q -N ""
+fi
+pub_ssh_key=$(cut -d " " -f "1,2" "${IMAGE_DIR}/ssh_testkey.pub")
 
 # Build an ISO image with baked ssh key
 # See https://www.system-rescue.org/scripts/sysrescue-customize/
 # We use the systemrescue ISO and their script for customizing it.
-pushd "${IMAGE_DIR}"
-wget -O sysrescue-customize https://gitlab.com/systemrescue/systemrescue-sources/-/raw/main/airootfs/usr/share/sysrescue/bin/sysrescue-customize?inline=false
-chmod +x sysrescue-customize
+if [[ ! -f "${IMAGE_DIR}/sysrescue-out.iso" ]];then
+    pushd "${IMAGE_DIR}"
+    wget -O sysrescue-customize https://gitlab.com/systemrescue/systemrescue-sources/-/raw/main/airootfs/usr/share/sysrescue/bin/sysrescue-customize?inline=false
+    chmod +x sysrescue-customize
 
-pub_ssh_key=$(cut -d " " -f "1,2" "ssh_testkey.pub")
-
-mkdir -p recipe/iso_add/sysrescue.d
-# Reference: https://www.system-rescue.org/manual/Configuring_SystemRescue/
-cat << EOF > recipe/iso_add/sysrescue.d/90-config.yaml
+    mkdir -p recipe/iso_add/sysrescue.d
+    # Reference: https://www.system-rescue.org/manual/Configuring_SystemRescue/
+    cat << EOF > recipe/iso_add/sysrescue.d/90-config.yaml
 ---
 global:
     nofirewall: true
@@ -149,9 +160,10 @@ sysconfig:
         "test@example.com": "${pub_ssh_key}"
 EOF
 
-./sysrescue-customize --auto --recipe-dir recipe --source systemrescue-11.00-amd64.iso --dest=sysrescue-out.iso
-export ISO_IMAGE_URL="http://${IP_ADDRESS}/sysrescue-out.iso"
-popd
+    ./sysrescue-customize --auto --recipe-dir recipe --source systemrescue-11.00-amd64.iso --dest=sysrescue-out.iso
+    export ISO_IMAGE_URL="http://${IP_ADDRESS}/sysrescue-out.iso"
+    popd
+fi
 
 # Generate credentials
 BMO_OVERLAYS=(
@@ -188,6 +200,12 @@ for overlay in "${IRONIC_OVERLAYS[@]}"; do
   envsubst < "${REPO_ROOT}/ironic-deployment/components/basic-auth/ironic-auth-config-tpl" > \
   "${overlay}/ironic-auth-config"
 done
+
+IRSO_IRONIC_AUTH_DIR="${REPO_ROOT}/test/e2e/data/ironic-standalone-operator/components/basic-auth"
+echo "${IRONIC_USERNAME}" > "${IRSO_IRONIC_AUTH_DIR}/ironic-username"
+echo "${IRONIC_PASSWORD}" > "${IRSO_IRONIC_AUTH_DIR}/ironic-password"
+
+sed -i "s|SSH_PUB_KEY_CONTENT|${pub_ssh_key}|" "${REPO_ROOT}"/test/e2e/data/ironic-standalone-operator/ironic/base/ironic.yaml
 
 # We need to gather artifacts/logs before exiting also if there are errors
 set +e
